Refactors All Exercises Except the mongodb
Refactors the exercises to break the exercises away from the tests, and
to use the new features in supposed.
