+++ exercises/110-functions/exercise.js
@@ -1,8 +1,7 @@
 'use strict';
 
-// >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
-// TODO: Refactor this code into a function
 
+// TODO: Refactor this code into a function named, `myCounter`
 let counter = 0;
 
 const counterOperations = {
@@ -14,30 +13,5 @@ const counterOperations = {
     }
 };
 
-// END TODO >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
-
-
-//////////////////////////////////////////////////////
-// TEST
-module.exports = (test) => test('110-functions', {
-  'myIncrementor, when it is reset and incrementCounter is called a given number of times': {
-      when: () => {
-          var reps = 5, i, lastRep, myIncrementor = myCounter();
-
-          myIncrementor.resetCounter();
-
-          for (i = 0; i < reps; i += 1) {
-              lastRep = myIncrementor.incrementCounter();
-          }
-
-          return {
-              expected: reps,
-              actual: lastRep
-          };
-      },
-      'it should increment the count the number of times requested': (t) => (err, result) => {
-          t.ifError(err);
-          t.equal(result.actual, result.expected);
-      }
-  }
-})
+// TODO: uncomment the next line:
+// module.exports = { myCounter };
