+++ exercises/120-iife/exercise.js
@@ -1,8 +1,7 @@
 'use strict';
 
-// >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
-// TODO: Refactor this code into an IIFE
 
+// TODO: Refactor this code into an IIFE named, `myIncrementor`
 let counter = 0;
 
 const counterOperations = {
@@ -14,27 +13,5 @@ const counterOperations = {
     }
 };
 
-// END TODO >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
-
-module.exports = (test) => test('120-iife', {
-  'myIncrementor, when it is reset and incrementCounter is called a given number of times': {
-      when: () => {
-          var reps = 5, i, lastRep;
-
-          myIncrementor.resetCounter();
-
-          for (i = 0; i < reps; i += 1) {
-              lastRep = myIncrementor.incrementCounter();
-          }
-
-          return {
-              expected: reps,
-              actual: lastRep
-          };
-      },
-      'it should increment the count the number of times requested': (t) => (err, result) => {
-          t.ifError(err);
-          t.equal(result.actual, result.expected);
-      }
-  }
-})
+// TODO: uncomment the next line:
+// module.exports = { myIncrementor };
