+++ exercises/322-ioc-containers/test.js
@@ -1,26 +1,29 @@
 'use strict';
 
-const test = require('supposed');
 const scope = require('./app.js');
 const http = require('http');
-var sutSingleton;
+let sutSingleton;
 
-module.exports = test('(SOLID::05-03-ioc-containers)', {
-    'when the app starts up': {
-        when: theServerStarts,
-        'it should return a server, db, logger, and port': appStarts,
-        'and I make a GET request for a product': {
-            when: iRequestAProduct,
-            'it should return a product': returnsAProduct,
-            'it should use the mock data connection': usesMockDb,
-            'it should use the mock logger': usesMockLogger
-        }
-    }
-}).then(() => {
+module.exports = (test, dependencies) => {
+  dependencies.teardown(() => {
     if (sutSingleton && sutSingleton.server) {
         sutSingleton.server.close();
     }
-});
+  })
+
+  return test('322-ioc-containers', {
+      'when the app starts up': {
+          when: theServerStarts,
+          'it should return a server, db, logger, and port': appStarts,
+          'and I make a GET request for a product': {
+              when: iRequestAProduct,
+              'it should return a product': returnsAProduct,
+              'it should use the mock data connection': usesMockDb,
+              'it should use the mock logger': usesMockLogger
+          }
+      }
+  });
+};
 
 function theServerStarts () {
     return new Promise((resolve, reject) => {
@@ -58,7 +61,7 @@ function iRequestAProduct () {
         theServerStarts().then(app => {
             http.get({
                 host: 'localhost',
-                port: 3000,
+                port: 3022,
                 path: '/products/42'
             }, function (res) {
                 // Continuously update stream with data
@@ -80,7 +83,7 @@ function iRequestAProduct () {
 function appStarts (t, err, app) {
     t.ifError(err);
     t.equal(typeof app.server, 'object');
-    t.equal(3000, 3000);
+    t.equal(3022, 3022);
     t.equal(typeof app.db, 'function');
     t.equal(typeof app.logger, 'object');
 }
