+++ exercises/index.js
@@ -1,6 +1,15 @@
-const suite = require('supposed').Suite({ name: 'heinz-95729-exercises'})
+const teardowns = []
+const suite = require('supposed').Suite({
+  name: 'heinz-95729-exercises',
+  inject: {
+    teardown: (teardown) => teardowns.push(teardown)
+  }
+})
 
 suite.runner({
   cwd: __dirname,
-  matchesNamingConvention: /.(exercise\.js)$/i,
+  matchesNamingConvention: /.(test\.js)$/i,
 }).run()
+  .then(() => {
+    teardowns.forEach((teardown) => teardown())
+  })
